requirements.txt
gradio>=4.44.0
google-generativeai
python-docx
PyMuPDF
pandas